requirements.txt
# Core AI Framework
crewai==0.30.11
crewai-tools==0.2.6

# LLM Providers (FREE options)
langchain==0.1.16
langchain-groq==0.1.3
langchain-community==0.0.38

# Vector Database (FREE)
chromadb==0.4.24

# Web Search (FREE)
duckduckgo-search==5.3.0

# UI Framework
streamlit==1.31.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.1
pydantic==2.7.0
requests==2.31.0
beautifulsoup4==4.12.3

# Development
pytest==8.2.0

black==24.1.1